api/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
psycopg2-binary==2.9.7
PyPDF2>=3.0.1
requests==2.31.0
python-dotenv==1.0.1
sentence-transformers>=2.2.2